Notes on the latest bug fix in kmem_manager.c:
It seems like the bug in question was caused by the fact that our flag definition where messed up. We are specifying to the kmem manager that we want KMEM_FLAG_WC for our mappings of the HBA region, but that flag is built up of the different flags KMEM_FLAG_NOCACHE, KMEM_FLAG_WRITETHROUGH and KMEM_FLAG_SPEC. Let it be known that we where also using the kmem_set_page_flags function at the time for mapping pages, which sets the PTE_PAT flag internally inside the actual page in memory, when KMEM_FLAG_SPEC is specified. Note that this is not a problem for the eventual pml1 page that we want to map and that also contains the actual physical address of the physical page we want to map, but this function (kmem_set_page_flags) was also used for the higher levels of the pagemap. These pages use different flag bits, and thus when kmem_set_page_flags tries to set PTE_PAT, what it is actually setting is PDE_HUGE_PAGE (both of these are bit 7 -> (1 << 7) as seen in pg.h).

This caused our mappings to confuse the cpu and it simply dies =/