Started on some process management with exceptions
modified interupts.c (yes, we need to fix the typo) a bit so that we now
detect if an exception is triggered by a process or by the kernel.

There seems to be some weird behaviour by our scheduler, because after
we kill the process in the pagefault handler, the scheduler tries to
scheduler the idle thread of the process! In this little dance it
swiches from and to different levels of privilige like so:
0: for the initial schedule of the process
3: for the execution of the process
0: for the handling of the exception, caused by the process
3: for trying to run the idle thread (in usermode)
0: for triggering another pagefault and ultimately getting stuck on a
   kernel_panic since it now realizes that this process needs to get
killed and it tries to murder the pagetable of this process, which we
don't yet know how to do =/