Officially fixed all known issues with elf
We have 'found' our bug in thread.c that caused a gpf when entering a
user process and it turned out that the mapping didn't really make stack
usage very happy. This might have something to do with the limits that
we set in the GDT/IDT, but idrk lol. It works (for) now so all is good
(for now)