Did some more syscall and LightEnv work
also, TODO: redo the way sockets and responses work. Right now responses
go through this really cumbersome mechanism with async pointers and
whatnot (which is very fancy indeed, but not what we're going for) so we
need to just implement some routines that any socket can call from
anywhere that just sends another packet to the sender of that packet. We
can keep the way we do it now, but instead of using weird async_ptrs, we
can just send another packet