Fixed memory leak in the khandle_map deletion
It seemed like the destruction of a khandle_map didn't go correctly and
that caused the process creation to halt after 3 launches. When
destructing a khandle_map, we deallocated a range of SMALL_PAGE_SIZE
every time, while the actual size of the range is way bigger, since it
needs to fit a default of 512 khandle objects, which are each 24 bytes
in size, inside the range. Thus this needs a total of ((512 * 24) >>
12) = 3 pages. We where deallocating 2 pages to little every time we
    created and destroyed a process XD