Fixed FAT32
Our bpb was fucked yay

we are now able to open and read from the FAT filesystem

We still need to create a valid check for the root filesystem on boot,
but this is step 0

Also did some refactoring of the vobj, vdir, ect. system. We now don't
attach a vobj to its parent node on creation, and the filesystem needs
to do that manually.
THERE IS STILL SOME AMBIGUITY HERE, WHICH IS NOT SO NICE =)