Added (it seems) working directory functionality in the oss and fs subystems
Every oss_node is now able to accept a dir_t helper object which can be administered by a filesystem driver. With this we're able to let filesystem drivers express
directories and userspace can then perform opperations on them (like removing/adding files).

The next step is to further generify this system so that we can have working dirs all around the oss tree.