More integration stuff =D Light loader and lightos are now one repo. I don't yet know how practical this will be, but we'll see.
This means the installation disk layout for lightos will now have a different format, since the bootloader and lightos now share the same partition and even the same filesystem. This might have to change in the future, when we want to give lightos flexible filesystem support. Right now it's all FAT32, since that is ez to implement and also convenient to boot off from UEFI.

This is a problem for future me =)
Thanks past me =(